Refactor config provider to not reload service incorrectly
The old reload notification didn't work when when the service provider
upstart. The user of consul_template_config will now need to now to
notify a reload of the consul-template service. This is a breaking
change, but much cleaner and simpler to reason about. It also will not
result in more than one reload (if :delayed is used).