Merge upstream to get systemd fix (#1)
* Reload systemd when template changes.

If the systemd template is changed, we need to ensure we perform a reload prior to restarting the consul-template service. Without this the changes will not be picked up.

* Change systemd service file to be non-executable

There's no reason for it to be, and systemd doesn't like it,
on Ubuntu 16.04.2 LTS at least:

dmesg:

systemd-sysv-generator[4385]: Configuration file
  /etc/systemd/system/consul-template.service is marked executable.
  Please remove executable permission bits. Proceeding anyway.

* update checksums for v0.18.0

* Added checksums for v0.19.0

* unpinned nssm version depedency

* updated suites with the correct centos-7.3

* adding custom consul/vault addr options

* break long line

* Fix checksums after merge

* Add 0.19.4 checksums and update default version

* Pin Test Kitchen to Chef 12

Will address Chef 13 compatibility soon, this just gets the test suite
working again.

* Updates test kitchen OS matrix

Removes Ubuntu 12.04 and Debian 7.11, adds Debian 9.3, updates Debian
8 and CentOS 6 and 7 minor versions

* Debian 9.3 box hasn't been created yet

* Removes support for supervisor

While I hate to remove functionality, the supervisor cookbook depends on
the deprecated python cookbook. This appears to be the only issue
preventing Chef 13 compatibility, so it has to go.

I'm not against adding this back in at some point in the future, if the
compatibility situation is fixed.

* Remote constraint on golang

* Update ruby version

* Rubocop fixes

* Foodcritic fixes

* Debian 8.10 box hasn't been created yet

* Allow specification of environment variables in init templates

* Removes installation from source for now

It looks like this has probably been broken for a while, so I'm
disabling it until I get a chance to rewrite it.

* Update CHANGELOG

* Mark default.json as sensitive resource

Closes #71

* Adds ExecStop to systemd service

Closes #72

* Update CHANGELOG

* Updates OS support in metadata

* v0.13.0 release